Bug: Espresso does not wait for data to be updated in UI
Since we use coroutines in a background thread to calculate UI
updates, espresso does not correctly wait for the UI to update in all
cases. (Espresso only waits for the main thread to be idle). This
causes random errors in our tests.

Implemented IdlingResource as recommended in Espresso docs. The issue
seems to be resolved now.

Unfortunately, this introduced some code into the application; the
sole purpose of which is to help out Espresso. :(